+++ include/boost/beast/http/empty_body.hpp
@@ -100,7 +100,7 @@ struct empty_body
     struct writer
     {
         using const_buffers_type =
-            boost::asio::null_buffers;
+            boost::asio::const_buffer;
 
         template<bool isRequest, class Fields>
         explicit