+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Version 145:
 * Rename some detail functions
 * Refactor basic_fields allocator internals
 * Refactor HTTP async read composed operations
+* null_buffers is deprecated
 
 --------------------------------------------------------------------------------